refactor(nextjs): restructure routing modules used in the handler
