+++ lib/presets/custom/next/compute/default/handler/routing/routes-matcher.js
@@ -1,18 +1,19 @@
 import { parse } from 'cookie';
 import {
+  runOrFetchBuildOutputItem,
+  getNextPhase,
+  isLocaleTrailingSlashRegex,
+} from './utils.js';
+import {
   parseAcceptLanguage,
   applyHeaders,
-  applyPCREMatches,
   applySearchParams,
-  hasField,
   isUrl,
-  matchPCRE,
-  isLocaleTrailingSlashRegex,
-} from './index.js';
-import { runOrFetchBuildOutputItem, getNextPhase } from './utils.js';
+} from './http.js';
+import { applyPCREMatches, matchPCRE } from './pcre.js';
+import { hasField } from './matcher.js';
 
-// eslint-disable-next-line import/prefer-default-export
-export class RoutesMatcher {
+class RoutesMatcher {
   /**
    * Creates a new instance of a request matcher.
    *
@@ -534,3 +535,6 @@ export class RoutesMatcher {
     return result;
   }
 }
+
+// eslint-disable-next-line import/prefer-default-export
+export { RoutesMatcher };
