+++ lib/presets/custom/next/compute/default/handler/routing/utils.js
@@ -1,10 +1,9 @@
-/* eslint-disable no-undef */
 import {
   applyHeaders,
   createRouteRequest,
   createMutableResponse,
   applySearchParams,
-} from './index.js';
+} from './http.js';
 
 /**
  * Gets the next phase of the routing process.
@@ -14,7 +13,7 @@ import {
  * @param phase Current phase of the routing process.
  * @returns Next phase of the routing process.
  */
-export function getNextPhase(phase) {
+function getNextPhase(phase) {
   switch (phase) {
     // `none` applied headers/redirects/middleware/`beforeFiles` rewrites. It checked non-dynamic routes and static assets.
     case 'none': {
@@ -52,7 +51,7 @@ export function getNextPhase(phase) {
  * @param match.searchParams
  * @returns Response object.
  */
-export async function runOrFetchBuildOutputItem(
+async function runOrFetchBuildOutputItem(
   item,
   { request, assetsFetcher, ctx },
   { path, searchParams },
@@ -105,3 +104,26 @@ export async function runOrFetchBuildOutputItem(
 
   return createMutableResponse(resp);
 }
+
+/**
+ * Checks if a source route's matcher uses the regex format for locales with a trailing slash, where
+ * the locales specified are known.
+ *
+ * Determines whether a matcher is in the format of `^//?(?:en|fr|nl)/(.*)`.
+ * @param src Source route `src` regex value.
+ * @param locales Known available locales.
+ * @returns Whether the source route matches the regex for a locale with a trailing slash.
+ */
+function isLocaleTrailingSlashRegex(src, locales) {
+  const prefix = '^//?(?:';
+  const suffix = ')/(.*)';
+
+  if (!src.startsWith(prefix) || !src.endsWith(suffix)) {
+    return false;
+  }
+
+  const foundLocales = src.slice(prefix.length, -suffix.length).split('|');
+  return foundLocales.every((locale) => locale in locales);
+}
+
+export { getNextPhase, runOrFetchBuildOutputItem, isLocaleTrailingSlashRegex };
