+++ lib/presets/custom/next/compute/default/prebuild/mapping/index.js
@@ -18,8 +18,8 @@ function isVcConfigValid(vcConfig) {
   const validRuntime = isEdge || isNode;
 
   // valid framework versions for runtimes
-  const validNextNodeVersions = ['12.3.1'];
-  const validNextEdgeVersions = ['12.3.1', '13.0.0'];
+  const validNextNodeVersions = ['12.3.1', '12.3.4'];
+  const validNextEdgeVersions = ['12.3.1', '12.3.4', '13.0.0'];
 
   let validVersion = false;
   let validEntrypoint = false;
