+++ lib/presets/custom/next/compute/handler.js
@@ -1,8 +1,10 @@
 /* eslint-disable */
 import {
-  adjustRequestForVercel,
   handleRequest,
 } from 'VULCAN_LIB_PATH/presets/custom/next/compute/default/handler/routing/index.js';
+import {
+  adjustRequestForVercel
+} from 'VULCAN_LIB_PATH/presets/custom/next/compute/default/handler/routing/http.js';
 
 const getStorageAsset = async (request) => {
   const VERSION_ID = AZION_VERSION_ID;
