+++ lib/presets/custom/next/compute/node/handler/index.js
@@ -2,7 +2,8 @@
 // .edge files are dynamically generated
 import { assets } from './.edge/next-build/statics.js';
 
-const createServer = require('./.edge/next-build/custom-server/index.js').default;
+const createServer =
+  require('./.edge/next-build/custom-server/index.js').default;
 
 globalThis.runNodeCustomServer = async function runNodeCustomServer(request) {
   const server = await createServer({
